requirements.txt
django-filter
Django==4.2.7
#django-mysql==4.12.0
#mysqlclient==2.2.4
djangorestframework==3.14.0
black==25.1.0

